Fix + add items

+++ src/space_2/MarketItemDetailView.cpp
@@ -150,7 +150,7 @@ void MarketItemDetailView::notifyItemChanged()
 		this->mTBTier.setFontColor(this->getItem()->getItemTier()->getColor());
 		this->mTBArianeType.setText("(" + this->getItem()->getItemType()->getAriane() + ")");
 		this->mTBDescription.setText(this->getItem()->getDescription());
-		this->mTBPrice.setText(Tools::formatNumber((int)this->getItem()->getPrice()) + " " + Resource::resource->getBundle()->getString("creditAb"));
+		this->mTBPrice.setText(Tools::formatNumber(this->getItem()->getPrice()) + " " + Resource::resource->getBundle()->getString("creditAb"));
 
 		this->updatePosition();
 		this->updateHeight();
